# openapi.yaml
openapi: 3.0.0
info:
  title: api
  version: 0.1.0
paths:
  /api/data/charts:
    get:
      tags:
      - Data
      description: Get prices
      operationId: get_charts
      responses:
        default:
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: uint32
                  minimum: 0.0
  /api/data/news:
    get:
      tags:
      - Data
      operationId: get_news
      parameters:
      - name: ticker
        in: query
        schema:
          type: string
          nullable: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsItem'
        '404':
          description: Unauthorized access
        '500':
          description: Internal server error
        '409':
          description: Email already registered
  /api/user/delete:
    post:
      tags:
      - User
      operationId: delete
      responses:
        '200':
          description: ''
          content:
            text/plain:
              schema:
                type: string
  /api/user/signup:
    post:
      tags:
      - User
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
        required: true
      responses:
        '200':
          description: ''
        '404':
          description: Unauthorized access
        '500':
          description: Internal server error
        '409':
          description: Email already registered
  /api/user/login:
    post:
      tags:
      - User
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
        required: true
      responses:
        '200':
          description: ''
        '404':
          description: Unauthorized access
        '500':
          description: Internal server error
        '409':
          description: Email already registered
  /api/data/market-overview:
    get:
      tags:
      - Data
      operationId: get_market_overview
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarketOverviewItem'
        '404':
          description: Unauthorized access
        '500':
          description: Internal server error
        '409':
          description: Email already registered
  /:
    get:
      tags:
      - Landing
      operationId: landing
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Unauthorized access
        '500':
          description: Internal server error
        '409':
          description: Email already registered
components:
  schemas:
    NewsItem:
      type: object
      required:
      - date
      - id
      - publisher
      - source
      - title
      properties:
        id:
          type: string
        title:
          type: string
        publisher:
          type: string
        source:
          type: string
        date:
          type: string
    UserCredentials:
      type: object
      required:
      - email
      - password_hash
      properties:
        email:
          type: string
        password_hash:
          type: string
    MarketOverviewItem:
      type: object
      required:
      - change
      - current_price
      - high
      - low
      - name
      - sector
      - short
      - symbol
      - volume
      properties:
        name:
          type: string
        short:
          type: string
        sector:
          type: string
        current_price:
          type: string
        change:
          type: number
          format: double
        high:
          type: number
          format: double
        low:
          type: number
          format: double
        symbol:
          type: string
        volume:
          type: integer
          format: uint64
          minimum: 0.0
        news_article:
          allOf:
          - $ref: '#/components/schemas/NewsItem'
          nullable: true